[DF-9967] add error messeges, fix local config bug (we create it on read)
Former-commit-id: bc930d8e974e669fbb3984da2e4b7831180def95

+++ Tools/Launcher/Classes/filemanager.cpp
@@ -1,4 +1,5 @@
 #include "filemanager.h"
+#include "errormessenger.h"
 #include <QDesktopServices>
 #include <QCoreApplication>
 #include <QDir>
@@ -104,6 +105,11 @@ bool CreateFileAndWriteData(const QString& filePath, const QByteArray& data)
 
 bool DeleteDirectory(const QString& path)
 {
+    if (path == "/" || path == "." || path == "..")
+    {
+        ErrorMessenger::ShowErrorMessage(ErrorMessenger::ERROR_PATH, "trying to remove wrong path! Aborted");
+        return false;
+    }
     QDir dir(path);
     return dir.removeRecursively();
 }
